+++ src/netcore/tests/task-runners/failure.ts
@@ -1,0 +1,9 @@
+import path = require('path');
+import models = require('../builders/request-model-builder');
+
+const taskPath = path.join(__dirname, '..\\..\\index.js');
+
+new models.RequestModel(taskPath)
+    .withFileNames('\n**\\AssemblyInfo.cs\n**\\NetCoreLib.csproj')
+    .withFailOnWarning(true)
+    .build();